requirements.txt
Flask
Fabric
cryptography
celery
redis
pywinrm